+++ loqusdb/commands/export.py
@@ -56,9 +56,13 @@ def export(ctx, outfile, variant_type, freq):
             if genome == GRCH37 and chrom in existing_chromosomes:
                 ordered_chromosomes.append(chrom)
                 existing_chromosomes.remove(chrom)
-            elif genome == GRCH38 and chrom[3:] in existing_chromosomes:
-                ordered_chromosomes.append(chrom)
-                existing_chromosomes.remove(chrom)
+            elif genome == GRCH38:
+                if chrom[3:] in existing_chromosomes:
+                    ordered_chromosomes.append(chrom[3:])
+                    existing_chromosomes.remove(chrom[3:])
+                elif chrom in existing_chromosomes:
+                    ordered_chromosomes.append(chrom)
+                    existing_chromosomes.remove(chrom)
     for chrom in existing_chromosomes:
         ordered_chromosomes.append(chrom)